third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-exception-source-url-expected.txt

CONSOLE ERROR: line 1: Uncaught ReferenceError: FAIL is not defined
Tests that when exception happens before inspector is open source url is correctly shown in console.

foo2.js:1 Uncaught ReferenceError: FAIL is not defined
    at foo (foo2.js:1:18)
    at onload (http://127.0.0.1:8000/inspector-enabled/console-exception-source-url.html:14:5)
    at onload (http://127.0.0.1:8000/inspector-enabled/console-exception-source-url.html:34:25)
TEST COMPLETE
